Breaking: Plugin system.
This creates a plugin system for code that is reusable but not suitable
for the framework. Plugins can tie into the server code, providing both
handlers and commands, as well as add additional make recipes and
additional server-side dependencies.

Authentication and uploading are both split into plugins and removed
from this repository, since they are not used by the rest of the
framework.

This commit includes some related cleanup, mostly renaming files.